chore: move map-related classes for better organization
Moved CurrentPosition and LocationNameProvider to the util/map directory for improved structure and clarity in the project.
