+++ app/src/main/java/fr/iut_rodez/pathpilot_android_client/util/map/CurrentPosition.java
@@ -1,4 +1,4 @@
-package fr.iut_rodez.pathpilot_android_client.map;
+package fr.iut_rodez.pathpilot_android_client.util.map;
 
 import static androidx.core.content.ContextCompat.checkSelfPermission;
